fix(lists): mobile view
This provides a nice mobile view for nav-finder.

There are some interesting numbers in here that are, as you might
notice, not a variable - this is because they are not set to a
specific common layout-based breakpoint, but to a point where the
content starts to look off. We can modify if we find them not great.