feat(lists): Add functionality to nav-finder
nav-finder needed to be far less content aware. It now has the ability
to be 2-4 elements long with the same structure, and it will cut text
with an ellipsis if something overruns.

A user can also use the `lvl-*` option for the instance where some text
can be hidden at a certain BP.