fix(forms): remove max default on form stepper
There should be no max default for the stepper component

BREAKING CHANGE: This could break tests down the line and any
assumptions made when the max value was not set. Originally, the max
defaulted to 99999. If a user had left the `max` attribute out of the
input element, with this change, there is no longer that default max to
fall back on