fix(iconography): icon-avatar -> avatar-icon
The `icon-avatar` class was being paired with `icon` in documentation,
and this was causing a conflict when both were being used together. So
this use-case has been removed from the docs.

Also, `icon-avatar` is being changed to `avatar-icon` so as to limit
confusion if `icon` and `icon-avatar` should be used together.

BREAKING CHANGE: Any case where icon-avatar is being used needs to be
changed to avatar-icon and if there is a sibling icon class, it needs to
be removed