feat(marketing): Create custom panel module
There are times when a panel just needs to be custom. This is where we
build those. The first of these panels is the module revolution panel.
It involves a big inline SVG. This big inline SVG will need a new way to
be shown in the code example, which means a likely change to PUI is
incoming.