feat(forms): add constraint validator
Form validation needed an overhaul. We don't have proper frontend
validation and we don't necessarily have the time to fold something in.
There should be a singular place to handle when/how errors show and they should
use the constraint validation api in order to do so.

You can opt-in to this functionality by using the `novalidate` attribute
on a form. This seems counterintuitive, but it will force custom
validation handling over what the browser does by itself.

BREAKING CHANGE: This changes the way link updater works, as it
separates the concerns of updating the link underneath and handling the
validation/error showing.