fix(forms): Link updater handles own errors
When using the link updater, we want to handle errors within the link
updater module itself. This incurs tech debt for us, as it requires
changing the way errors look in multiple places now.

There are potential fixes for that, though, one of which is to make
adding and removing errors a function of the validator itself.

[Fixes #126237181]