fix(forms): Remove absolute positioning on errors
Absolute positioning was a fun hack that allowed for a really nice
display, but made this thing unworkable with other stacked errors.

We've moved ourselves to a js-based transition and block-positioned
elements in order to get away from that.

This will require screenshots of both the existing and the new version
in order to compare/contrast make sure they look exactly the same so
this won't break current views